Move structs to models & follow naming convention
These probably shouldn't have lived in these modules, they were just
where I wrote them originally. They should both live in `models` along
with other input payloads.
